+++ src/ArangoDB.Client.Test/Mock/ArangoDatabaseMock.cs
@@ -2,6 +2,7 @@
 using Moq;
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
@@ -56,7 +57,7 @@ namespace ArangoDB.Client.Test.Mock
         public ArangoDatabaseMock SendCommand(string json, HttpStatusCode? statusCode = HttpStatusCode.OK)
         {
             mockDB.Setup(x => x.Connection
-                .SendCommandAsync(It.IsAny<HttpMethod>(), It.IsAny<Uri>(), It.IsAny<object>(), It.IsAny<NetworkCredential>()))
+                .SendCommandAsync(It.IsAny<HttpMethod>(), It.IsAny<Uri>(), It.IsAny<object>(),It.IsAny<Func<StreamWriter,Task>>(), It.IsAny<NetworkCredential>()))
                 .ReturnsAsync(new HttpResponseMessage() { StatusCode = statusCode.Value, Content = new StringContent(json) });
 
             return this;
@@ -65,7 +66,7 @@ namespace ArangoDB.Client.Test.Mock
         public ArangoDatabaseMock SendCommandSequence(IList<string> json, HttpStatusCode? statusCode = HttpStatusCode.OK)
         {
             var setup = mockDB.SetupSequence(x => x.Connection
-                .SendCommandAsync(It.IsAny<HttpMethod>(), It.IsAny<Uri>(), It.IsAny<object>(), It.IsAny<NetworkCredential>()));
+                .SendCommandAsync(It.IsAny<HttpMethod>(), It.IsAny<Uri>(), It.IsAny<object>(), It.IsAny<Func<StreamWriter, Task>>(), It.IsAny<NetworkCredential>()));
 
 
             foreach(var j in json)