+++ src/ArangoDB.Client/Http/IHttpConnection.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
@@ -10,6 +11,6 @@ namespace ArangoDB.Client.Http
 {
     public interface IHttpConnection
     {
-        Task<HttpResponseMessage> SendCommandAsync(HttpMethod method, Uri uri, object data, NetworkCredential credential);
+        Task<HttpResponseMessage> SendCommandAsync(HttpMethod method, Uri uri, object data, Func<StreamWriter, Task> onStreamReady, NetworkCredential credential);
     }
 }